+++ packages/urls/urls.js
@@ -87,5 +87,14 @@ ggUrls.saveGameRule =function(gameRuleSlug) {
 };
 
 ggUrls.user =function(username) {
-  return '/u/'+username;
+  return '/u' + ( username ? ( '/' + username ) : '' );
+};
+ggUrls.userProfile =function(username) {
+  return ggUrls.user(username) + '/?nav=profile';
+};
+ggUrls.userGameAwards =function(username) {
+  return ggUrls.user(username) + '/?nav=game-awards';
+};
+ggUrls.userGameTimeline =function(username) {
+  return ggUrls.user(username) + '/?nav=game-timeline';
 };