+++ users/user-game-awards/user-game-awards.js
@@ -1,0 +1,13 @@
+if(Meteor.isClient) {
+  // Template.userGameAwards.created =function() {
+  // };
+
+  Template.userGameAwards.helpers({
+    data: function() {
+      var user =this.user;
+      return {
+        user: user
+      }
+    }
+  });
+}