+++ users/user-game-timeline/user-game-timeline.js
@@ -1,0 +1,10 @@
+if(Meteor.isClient) {
+  Template.userGameTimeline.helpers({
+    data: function() {
+      var user =this.user;
+      return {
+        user: user
+      }
+    }
+  });
+}